# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0",
    "title": "API Veículos",
    "description": "API do projeto de fabrica de software WEBCARS"
  },
  "servers": [
    {
      "title": "Web cars",
      "url": "https://webcars.dev.vilhena.ifro.edu.br/api/"
    }
  ],
...
  "paths": {
    "/carro/imagem/{idImagem}": {
      "get": {
        "tags": [
          "Carro"
        ],
        "description": "Retorna a imagem de um carro pelo ID da imagem",
        "parameters": [
          {
            "name": "idImagem",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da imagem do carro"
          }
        ],
        "responses": {
          "200": {
            "description": "Imagem do carro encontrada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "string"
                    },
                    "format": {
                      "type": "string",
                      "example": "binary"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "string"
                    },
                    "format": {
                      "type": "string",
                      "example": "binary"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Imagem não encontrada"
          },
          "500": {
            "description": "Erro interno ao buscar a imagem"
          }
        }
      }
    },
...
    "/aro/": {
      "post": {
        "tags": [
          "Aro"
        ],
        "description": "Cadastra um aro",
        "parameters": [
          {
            "name": "aro",
            "in": "body",
            "description": "Dados do aro",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Aro"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Aro"
        ],
        "description": "Retorna todos os aros e caso tenha o parâmetro nome, lista o câmbio pelo nome",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "description": "Nome do aro para filtrar",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/cambio/": {
      "post": {
        "tags": [
          "Câmbio"
        ],
        "description": "Cadastra um câmbio",
        "parameters": [
          {
            "name": "cambio",
            "in": "body",
            "description": "Dados do câmbio",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Cambio"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Câmbio"
        ],
        "description": "Lista todos os câmbios e caso tenha o parâmetro nome, lista o câmbio pelo nome",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "description": "Nome do câmbio",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/categoria/": {
      "post": {
        "tags": [
          "Categoria"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Criado"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Categoria"
        ],
        "description": "",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/combustivel/": {
      "post": {
        "tags": [
          "Combustível"
        ],
        "description": "Cadastra um combustível",
        "parameters": [
          {
            "name": "combustivel",
            "in": "body",
            "description": "Dados do combustível",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Combustivel"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Combustível"
        ],
        "description": "Lista todos os combustíveis",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "description": "Nome do combustível",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/cor/": {
      "post": {
        "tags": [
          "Cor"
        ],
        "description": "Cadastra uma nova cor",
        "parameters": [
          {
            "name": "cor",
            "in": "body",
            "description": "Dados da cor",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Cor"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Cor"
        ],
        "description": "Lista todas as cores",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "description": "Nome da cor para busca",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/marca/": {
      "post": {
        "tags": [
          "Marca"
        ],
        "description": "Cadastra uma marca",
        "parameters": [
          {
            "name": "marca",
            "in": "body",
            "description": "Dados da marca",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Marca"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Marca"
        ],
        "description": "Lista todas as marcas ou busca por nome",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "description": "Nome da marca para busca",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/modelo/": {
      "post": {
        "tags": [
          "Modelo"
        ],
        "description": "Cadastra um modelo",
        "parameters": [
          {
            "name": "modelo",
            "in": "body",
            "description": "Dados do modelo",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Modelo"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  },
                  "id_marca": {
                    "example": "any"
                  },
                  "id_categoria": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Modelo"
        ],
        "description": "Lista todos os modelos ou busca por nome",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "description": "Nome do modelo",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/modelo/marca/": {
      "get": {
        "tags": [
          "Modelo"
        ],
        "description": "Lista todos os modelos ou busca por nome da marca",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "description": "Nome da marca",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/modelo/categoria/": {
      "get": {
        "tags": [
          "Modelo"
        ],
        "description": "Lista todos os modelos ou busca por nome da categoria",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "description": "Nome da categoria",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/carro/": {
      "get": {
        "tags": [
          "Carro"
        ],
        "description": "Lista todos os carros ou busca por nome",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "description": "Nome do carro",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/filtro-alerta/": {
      "get": {
        "tags": [
          "FiltroAlerta"
        ],
        "description": "Lista filtros de alerta",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "FiltroAlerta"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Criado"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FiltroAlerta"
              }
            }
          }
        }
      }
    },
...
    "/cliente/": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "description": "Cadastra um novo cliente",
        "parameters": [
          {
            "name": "cliente",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Cliente"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado"
          }
        }
      },
      "get": {
        "tags": [
          "Cliente"
        ],
        "description": "Retorna todos os clientes",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/concessionaria/": {
      "post": {
        "tags": [
          "Concessionaria"
        ],
        "description": "Cadastra uma nova concessionária",
        "parameters": [
          {
            "name": "concessionaria",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Concessionaria"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado"
          }
        }
      },
      "get": {
        "tags": [
          "Concessionaria"
        ],
        "description": "Retorna todas as concessionárias",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FiltroAlerta": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "nome_filtroAlerta": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Filtro SUV Premium"
                  }
                }
              },
              "ano": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 2022
                  }
                }
              },
              "condicao": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Semi-novo"
                  }
                }
              },
              "ipva_pago": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  },
                  "example": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "blindagem": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  },
                  "example": {
                    "type": "boolean",
                    "example": false
                  }
                }
              },
              "data_ipva": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date"
                  },
                  "example": {
                    "type": "string",
                    "example": "2023-01-15"
                  }
                }
              },
              "data_compra": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date"
                  },
                  "example": {
                    "type": "string",
                    "example": "2023-01-01"
                  }
                }
              },
              "valor_maximo": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  },
                  "example": {
                    "type": "number",
                    "example": 150000
                  }
                }
              },
              "valor_minimo": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  },
                  "example": {
                    "type": "number",
                    "example": 80000
                  }
                }
              },
              "cor_id_cor": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 3
                  }
                }
              },
              "cambio_id_cambio": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "aro_id_aro": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 2
                  }
                }
              },
              "categoria_id_categoria": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "marca_id_marca": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 5
                  }
                }
              },
              "combustivel_id_combustivel": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 2
                  }
                }
              },
              "cliente_id_cliente": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "modelo_id_modelo": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "FiltroAlerta"
        }
      }
    }
  }
}